renaming folder ViewModel to ViewModels
