--- storage_app/ViewModels/MainWindowViewModel.cs
@@ -1,25 +1,0 @@
-﻿//using storage_app.Models;
-//using System;
-//using System.Collections.Generic;
-//using System.Linq;
-//using System.Text;
-//using System.Threading.Tasks;
-
-//namespace storage_app.ViewModels
-//{
-//    public class MainWindowViewModel : ViewModelBase
-//    {
-//        public ICollection<Product> Products { get; set; }
-//        public MainWindowViewModel()
-//        {
-//            Products = new List<Product>();
-
-//            Product _product1 = new();
-//            Product _product2 = new();
-//            _product1.Description = "O Hobbit";
-//            _product2.Description = "O Silmarillion";
-//            Products.Add(_product1);
-//            Products.Add(_product2);
-//        }
-//    }
-//}
